backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
cryptography>=43.0.0
pydantic>=2.7.0
pytest>=8.2.0
requests>=2.32.0
python-dotenv>=1.0.1
